requirements.txt
pillow>=9.0.0
onnxruntime>=1.12.0
huggingface-hub
gradio